almost complete Aho-Corasick string matcher
This is a very simple implementation that, for the moment, lacks support for
Unicode.

As this is a piece of library code, errors need to be treated carefully.
At the moment, I do not know how to do this.

The interface needs a bit of a cleanup, "generic lists" are trivial at the moment.
Need to talk with MJ about the use of hashmaps in linked lists.